abstract BaseStateMachine
